split out PartialUpdateInfo in a new file
