+++ be/src/olap/rowset_builder.h
@@ -32,6 +32,7 @@
 #include "common/status.h"
 #include "olap/delta_writer_context.h"
 #include "olap/olap_common.h"
+#include "olap/partial_update_info.h"
 #include "olap/rowset/rowset.h"
 #include "olap/tablet.h"
 #include "olap/tablet_meta.h"
